My modifications (May 2, 1995) for very slow terminals.
It is hard to believe today, but at 2400 bps, it takes 8 seconds
to repaint an 80x24 text window. This was a modification to do
that repainting as lazily as I could manage, so that if you knew
what you were doing and could edit or scroll faster than the
screen could repaint, it wouldn't waste time drawing the parts
that were already out of date.